Year numberrs removed from Copyright notices, xls_import dummy file added
