+++ pyspread/src/actions/test/test1.pys.sig
@@ -1,7 +1,7 @@
 -----BEGIN PGP SIGNATURE-----
 Version: GnuPG v1.4.12 (GNU/Linux)
 
-iF4EABEIAAYFAlGY/RcACgkQOipFccilpPPrHAD/VM9dRcqaB362Y83w2zeLnEdy
-VRFCwfuI2iY5/GuzZ7cA/A53vmD/EcRBrPgm3K0nlQ7GVNkoMEeQWNx9oLiQR/IX
-=Up+3
+iF4EABEIAAYFAlGZaZEACgkQOipFccilpPOevwEApZFKSSVqCPcMZLmwutfsVBU5
+kn3Ee1Dg98OSItV64m0BAMt9yxF5+5FmurcKQRnco0o1+nAmuX//vDRzJxdxlxpj
+=gZfT
 -----END PGP SIGNATURE-----
