+++ pyspread/src/actions/test/test4.pys.sig
@@ -1,7 +1,7 @@
 -----BEGIN PGP SIGNATURE-----
 Version: GnuPG v1.4.12 (GNU/Linux)
 
-iF4EABEIAAYFAlGY/RcACgkQOipFccilpPMwMAEAgDMS2PWRPq2XAgXROU/wc0E/
-wXqn05f4QIFGmMI7U+MBAMo5L6jBtcre/v/qzN4E9CiYpyeKy7oUaoCvTMBmsEA7
-=V3Vf
+iF4EABEIAAYFAlGZaZEACgkQOipFccilpPMVsgD9GJf8DPqCwNfBGXSKVW01FY0V
+cDFiV9uoDFicYbSD9qcA/iLzCSAYqKoqauKRqJE8du596jPxNTDeUTKA6Azx+t/6
+=WT2R
 -----END PGP SIGNATURE-----
