+++ pyspread/src/lib/test/test1.pys.sig
@@ -1,7 +1,7 @@
 -----BEGIN PGP SIGNATURE-----
 Version: GnuPG v1.4.12 (GNU/Linux)
 
-iF4EABEIAAYFAlGY/UQACgkQOipFccilpPOZ2gD/bQhcOm3fdgbtsxqX8nl+M9kw
-W+7ZLUTnFheuPpFe3qEA/2Ero9hLByJ6Md8dmAQuDWi5orvCbaVzX22ywIc8XnPV
-=xSnQ
+iF4EABEIAAYFAlGZabsACgkQOipFccilpPOvLAEAsb9jCED9M85OcPCJjjbngy18
+mA1BZDViI52H0U3vcEAA/Rg1cZt9QH1X35Yw6jLFVXzhmgQSkeNS033iZma5geRL
+=3N7f
 -----END PGP SIGNATURE-----
